Create method: //   FNullText := RsNoneCaption;  XXX Don't do this unless you also set the 'default' specifier in the property declaration above! Causes problems. -WP
The property specifier shows default string of '' (nothing), but inside the constructor, it is set to NOne, this means that at design time, the attempt to set a blank NullText property will be stomped at runtime. If we want a (none) defualt, then we must put default '(None)' in the property declaration.