NEW: Unique Y Axis Labels.
If the Y Axis is from 0 to 5 for example, and you had 10 Y divisions, you'd formerly get pairs of identical looking Y axis range labels like this: 0,0,1,1,2,2,3,3,4,4,5,5  Instead, we automatically add a decimal place to the Y Axis label until all labels are unique. So if you set 1 decimal place up in the chart it would always show it, need it or not, but if you set up 0 decimal places, it will now automaticaly give you just enough decimal places to give unique Y axis labels.