Fixed Mantis #0002126: TJvCalcEdit Button-width not saved
By republishing the ShowButton property, from the TJvCustomComboEdit ancestor class, the problem appears to be solved however a minor quirk remains.  When setting the ButtonWidth to 0 the ShowButton will automatically be set to False and the ButtonWidth property is reset again back to it's old property.  Not a big deal since the ShowButton property takes precedence.