I added one function and changed an other.
I overloaded the procedure AddFromStream, with a function returning the index of the object added, and added a new parameter aTitle. This one is needed to fix the problem that only ONE Thumbnail can be added from a stream.

I changed the procedure AddFromFile to a function returning the index of the object added.

No changes to sources using these procedures are needed