Added ForceDecimalSeparatorInput : when true and the user presses the VK_DECIMAL key, the key read in KeyPress will always be replaced by the value of DecimalSeparator.
This is made to overcome the problem where some keyboard layouts send "." instead of the decimal separator when using the decimal key on the numerical keypad.
The most commonly encountered layout with this behavior is the French AZERTY one.
Note that this property will be set automatically to True by the constructor when the conversion of VK_DECIMAL into a character does not return the DecimalSeparator value.