Mantis 0005035: TJvFormStorage not restore TPageControl.ActivePage
I've changed the handling of the TJvPropertyStorage.ReadProperty and TJvPropertyStorage.WriteProperty procedures. Now component properties are no longer stored with all the properties. only the name is stored. When reading the stored name the component is searched in the list of all components.

Someone else should have a look on this changes.