char conversion + Event init
FEATURE_SAVE_CONVERT_STRINGS

This feature uses a save way to convert the strings given by the Windows API.
It fixes a case, where the device can not be opened, because the (converted)
path is not valid.

TJvHidDeviceController.CreateInit() added

You need the OnDeviceCreateError event handler for catching and handling
of device errors.
The constructor will start the search for devices, you could get device
errors in this search, which will abort the search if no event method
is given.