Partial fix for Mantis 6457: TJvSpinEdit handles minus sign incorrectly
Rather than altering FNegative handling (the fix for #4572), it seems best
to check for '-' during keyboard input and discard it if it's invalid.

Disallow '+'; there seems to be no benefit to allowing it, since it's
discarded whenever the value is reformatted.

The existing code tries to allow scientific notation. I've tried to preserve
this, although I couldn't get scientific notation to work at all in my
testing (even before I made changes).