Merged #42: Small Docking library changes
https://github.com/project-jedi/jvcl/pull/42/commits/7a35bc1f011b1c57795175c42b9635b4753f977e

A) A text can be shown while the form is dragged or about to be docked
by using the events from the TjvDockStyle component (typically
OnFormGetSiteInfo)

B) In the TJvDockDragDockObject.MouseMsg proc, the TargetControl is
TJvDockPanel was commented out. This means when a TJvDockServer
component was on the target form the CheckIsDockable event from the
dragged windows was never called. By removing the comments it seems to
work correctly. Don't know why this part was commented out...