When changing Form options like 'BorderStyle' and 'BorderButtons' the CaptionButton position/rect needs to be recalculated.
Instead of letting the component automaticly figure out these changes and the fact that these things need to be changed in code, I thought is would be best to let the user notify the Button position/rect recalculationneed using the 'ResetButton' method.

I've updated the demo to show how to use this!