Add the ability to create a 'lookup like' picklist, even when FieldKind <> fkLookup. - this allows you to have different lookuplist for one field, based on another field - I can create a demo if wanted...
NEW FUNCTIONALITY:

Basicly the lookup information (based on TField.FieldKind = fkLookup) is wrapped by a new routine (GetColumnLookupInfo), which gets the lookup information from the column field.

In addition you can:
- hook the new 'OnGetColumnLookupInfo' event to specify the LookupInfo information
- override the GetColumnLookupInfo to specify the LookupInfo information

To ensure the picklist functionality is used 'GetEditStyle' is overriden.
As is 'CanEditModify', so the inplaceeditor will act like a 'lookuplist' (like fkLookup)