Drag&Drop support for the "New Outlook"
The "New Outlook" (WebApp) doesn't support synchronous OLE Drag&Drop and tells that the CF_HDROP format is available but when requesting the contents in OnDragAccept, OnDragEnter or OnDragOver, it fails with a DV_E_FORMATETC error code. So the GetFilenames/GetFilenamesW methods fail with an empty list.

With this patch the new TJvDropTarget.IsAsyncHDrop method can be used to find out if a HDROP format with asynchronous behavior is available. This method can be used in OnDragAccept, OnDragEnter and OnDragOver instead of querying the contents. The actual data becomes available in OnDragDrop when calling GetFilenames/GetFilenamesW. For the "New Outlook" the download that is started when calling GetFilenames/GetFilenamesW may have to be permitted in the WebApp window and the GetFilenames methods wait for that permission.