fix: Replaced usage of TextToShortCut(RsDefaultNoDateShortcut) with the constant integer value in TJvDatePickerEdit.
Not only is TextToShortCut(RsDefaultNoDateShortcut) slow but it also only works if the OS's language matches the language of the RsDefaultNoDateShortcut string. The constant doesn't need a translation for "Alt+Del" and so it doesn't fail depending on the used language.

The resourcestring RsDefaultNoDateShortcut was not removed, because there may be code out there that uses it.